// src/pages/Profile/types.ts
export interface ProfileData {
  firstname: string;
  lastname: string;
  birthdate: string;
  phone: string;
  address: string;
  postalcode: string;
  city: string;
  ffglicense: string;
  email: string;
  emergencycontact: string;
  emergencyphone: string;
  golfindex: number | null;
  birthplace: string;
}